[NFC][Codegen] Move distribution pattern infrastructure to VectorExt/Transforms
Move DistributionPattern, VectorLayoutOptions, DistributionSignature,
and related signature utility functions from Codegen/Common/GPU/ into
Codegen/Dialect/VectorExt/Transforms/DistributionPatterns.{h,cpp}.

The concrete NestedLayout distribution patterns remain in Common/GPU/
to avoid pulling Codegen-level dependencies into the VectorExt dialect
layer.

This change allows us to define distribution patterns outside of
Common/GPU without having a dependency on Common/GPU. This allows us to
define new patterns by only taking a dependency on VectorExtTransforms.
