+++ compiler/src/iree/compiler/Codegen/Common/GPU/GPUPatterns.h
@@ -35,10 +35,6 @@ void populateVectorLayoutCanonicalizations(RewritePatternSet &patterns);
 
 void populateGPUDistributionPatterns(RewritePatternSet &patterns);
 
-void populateGPUDistributeNestedLayoutAttrPatterns(
-    RewritePatternSet &patterns, Value threadId, int64_t subgroupSize,
-    ArrayRef<int64_t> workgroupSize, int64_t maxBitsPerShuffle = 32);
-
 } // namespace mlir::iree_compiler
 
 #endif // IREE_COMPILER_CODEGEN_COMMON_GPUPATTERNS_H_
