+++ compiler/src/iree/compiler/Codegen/Common/GPU/GPUVectorDistribution.h
@@ -7,159 +7,9 @@
 #ifndef IREE_COMPILER_CODEGEN_COMMON_GPU_VECTOR_DISTRIBUTION_H_
 #define IREE_COMPILER_CODEGEN_COMMON_GPU_VECTOR_DISTRIBUTION_H_
 
-#include "iree/compiler/Codegen/Dialect/VectorExt/IR/VectorExtOps.h"
-#include "llvm/Support/Debug.h"
-#include "mlir/Dialect/GPU/IR/GPUDialect.h"
-#include "mlir/Dialect/Vector/IR/VectorOps.h"
-#include "mlir/Interfaces/FunctionInterfaces.h"
+#include "iree/compiler/Codegen/Dialect/VectorExt/Transforms/DistributionPatterns.h"
 
 namespace mlir::iree_compiler {
-
-using IREE::VectorExt::VectorLayoutInterface;
-
-/// A signature describing the layout for each value of vector type which is
-/// an operand or result of this operation.
-///
-/// Two operands may be the same value, but since each value can only have
-/// one layout, we only need to keep track of the value, not the two operands
-/// separately.
-using DistributionSignature =
-    DenseMap<TypedValue<VectorType>, VectorLayoutInterface>;
-
-struct DistributionPattern : RewritePattern {
-  using RewritePattern::RewritePattern;
-
-  /// Lookup the distributed value for the given SIMD value. If the value
-  /// was not distributed yet, wrap it in a ToSIMTOp.
-  TypedValue<VectorType> getDistributed(RewriterBase &rewriter,
-                                        TypedValue<VectorType> value,
-                                        VectorLayoutInterface layout) const;
-
-  /// Get the distributed values that could replace the op
-  SmallVector<Value> getOpDistributedReplacements(RewriterBase &rewriter,
-                                                  Operation *op,
-                                                  ValueRange values) const;
-
-  /// Replace an op with its distributed replacement values.
-  void replaceOpWithDistributedValues(RewriterBase &rewriter, Operation *op,
-                                      ValueRange values) const;
-
-  /// Get the signature for the given operation.
-  std::optional<DistributionSignature> getOpSignature(Operation *op) const;
-
-protected:
-  // Sets new layout/signature for op, and mark it for redistribution.
-  // When "vector layout storage" and "vector layout redistribution"
-  // is defined, VectorDistributionRewriter would add it to worklist
-  // of operations to be distributed.
-  void setSignatureForRedistribution(
-      RewriterBase &rewriter, Operation *op,
-      ArrayRef<VectorLayoutInterface> inputLayouts,
-      ArrayRef<VectorLayoutInterface> outputLayouts) const;
-
-  LogicalResult replaceParentMask(PatternRewriter &rewriter,
-                                  vector::MaskOp) const;
-};
-
-template <typename SourceOp>
-struct OpDistributionPattern : DistributionPattern {
-  OpDistributionPattern(MLIRContext *context, PatternBenefit benefit = 1)
-      : DistributionPattern(SourceOp::getOperationName(), benefit, context) {}
-
-  virtual LogicalResult matchAndRewrite(SourceOp op,
-                                        DistributionSignature &opSignature,
-                                        PatternRewriter &rewriter) const = 0;
-
-  LogicalResult matchAndRewrite(Operation *op,
-                                PatternRewriter &rewriter) const final {
-    std::optional<DistributionSignature> opSignature = getOpSignature(op);
-    if (!opSignature) {
-      return failure();
-    }
-    return matchAndRewrite(cast<SourceOp>(op), *opSignature, rewriter);
-  }
-};
-
-template <typename SourceOp>
-struct MaskedOpDistributionPattern : DistributionPattern {
-  MaskedOpDistributionPattern(MLIRContext *context, PatternBenefit benefit = 1)
-      : DistributionPattern(SourceOp::getOperationName(), benefit, context) {}
-
-  virtual LogicalResult
-  matchAndRewrite(SourceOp op, DistributionSignature &opSignature,
-                  vector::MaskOp maskOp,
-                  std::optional<DistributionSignature> &maskSignature,
-                  PatternRewriter &rewriter) const = 0;
-
-  LogicalResult matchAndRewrite(Operation *op,
-                                PatternRewriter &rewriter) const final {
-    std::optional<DistributionSignature> opSignature = getOpSignature(op);
-    if (!opSignature) {
-      return failure();
-    }
-    auto maskOp = op->getParentOfType<vector::MaskOp>();
-    std::optional<DistributionSignature> maskSignature;
-    if (maskOp) {
-      maskSignature = getOpSignature(maskOp);
-      if (!maskSignature) {
-        return failure();
-      }
-    }
-    LogicalResult result = matchAndRewrite(cast<SourceOp>(op), *opSignature,
-                                           maskOp, maskSignature, rewriter);
-    if (failed(result)) {
-      return failure();
-    }
-    if (maskOp) {
-      return replaceParentMask(rewriter, maskOp);
-    }
-    return success();
-  }
-};
-
-template <template <typename> class TraitType>
-class OpTraitDistributionPattern : public DistributionPattern {
-public:
-  OpTraitDistributionPattern(MLIRContext *context, PatternBenefit benefit = 1)
-      : DistributionPattern(Pattern::MatchTraitOpTypeTag(),
-                            TypeID::get<TraitType>(), benefit, context) {}
-
-  virtual LogicalResult matchAndRewrite(Operation *op,
-                                        DistributionSignature &opSignature,
-                                        PatternRewriter &rewriter) const = 0;
-
-  LogicalResult matchAndRewrite(Operation *op,
-                                PatternRewriter &rewriter) const final {
-    std::optional<DistributionSignature> opSignature = getOpSignature(op);
-    if (!opSignature) {
-      return failure();
-    }
-    return matchAndRewrite(op, *opSignature, rewriter);
-  }
-};
-
-/// Options to control how the layout analysis is initialised for vector
-/// distribution.
-class VectorLayoutOptions {
-public:
-  VectorLayoutOptions(Operation *root) : root(root), fullConversion(true) {
-    assert(root && "root operation must be non-null");
-  }
-  VectorLayoutOptions(Operation *root, bool fullConversion)
-      : root(root), fullConversion(fullConversion) {
-    assert(root && "root operation must be non-null");
-  }
-
-  virtual ~VectorLayoutOptions() = default;
-
-  bool verifyConversion() const { return fullConversion; }
-
-  virtual VectorLayoutInterface getDefaultLayout(VectorType type) const = 0;
-
-protected:
-  Operation *root;
-  bool fullConversion = true;
-}; // namespace iree_compiler
 
 /// Distribute vector operations in the IR rooted at `root`.
 ///
@@ -173,9 +23,9 @@ protected:
 ///   - Run a global analysis to determine how to distribute rest of the vector
 ///     values keeping the initial anchors in mind.
 ///   - Use the analysis information to distribute each operation.
-LogicalResult distributeVectorOps(Operation *root,
-                                  RewritePatternSet &distributionPatterns,
-                                  VectorLayoutOptions &options);
+LogicalResult
+distributeVectorOps(Operation *root, RewritePatternSet &distributionPatterns,
+                    IREE::VectorExt::VectorLayoutOptions &options);
 
 } // namespace mlir::iree_compiler
