+++ compiler/src/iree/compiler/Codegen/Dialect/VectorExt/Transforms/BUILD.bazel
@@ -36,6 +36,7 @@ iree_compiler_cc_library(
     name = "VectorExtTransforms",
     srcs = [
         "BufferizationInterfaces.cpp",
+        "DistributionPatterns.cpp",
         "LowerTransferGatherOps.cpp",
         "Passes.cpp",
         "VectorExtFoldUnitExtentDims.cpp",
@@ -43,6 +44,7 @@ iree_compiler_cc_library(
     ],
     hdrs = [
         "BufferizationInterfaces.h",
+        "DistributionPatterns.h",
         "Passes.h",
         "Passes.h.inc",
         "Transforms.h",
